✨ Rename model files
