+++ server/src/client/src/apis/note.api.ts
@@ -1,4 +1,4 @@
-import type { Note } from '~/models/Note';
+import type { Note } from '~/models/note.model';
 import { graphQuery } from '~/modules/graph-query';
 
 export interface FetchNotesParams {
