+++ server/src/client/src/components/shared/Callout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Callout';
