+++ server/src/client/src/components/shared/index.ts
@@ -1,5 +1,6 @@
 export { default as Badge } from './Badge';
 export { default as Button } from './Button';
+export { default as Callout } from './Callout';
 export { default as Container } from './Container';
 export { default as Dropdown } from './Dropdown';
 export { default as FallbackRender } from './FallbackRender';
