+++ server/src/client/src/models/image.model.ts
