+++ server/src/client/src/models/note.model.ts
